Adapt Executor interface to real-time reading of messages
The telethon component will start one OS process for each chat that is
monitored. This process is a Python process running Telethon receiving
NewMessage updates from Telegram for this specific chat.

So, whenever the user adds a new peer in the Monitor class, the
startPeer() method is called, which will cause a new Telethon process to
be spun up for this chat. When the user removes a peer from the Monitor
class, the stopPeer() method is called, which causes the Telethon
process to be killed.

When the user calls the high-level start() and pause() methods of the
Monitor class, the following happens:
- start() calls startPeer() for each peer in the Monitor class
- pause() calls stopPeer() for each peer in the Monitor class
In the case of pause(), the list of peers in the Monitor class remains
unchanged (so that execution can be resumed with start() without the
user having to enter all the peers again).