Bottom-up dependencies via factories
When classes in lower components need to access the Monitor class in
this component, they do so via a factory. This factory gets and returns
the Monitor instance via a static get() method in the Monitor class.

The factory has an interface and a concrete implementation. The concrete
implementation goes together with the concrete implementation of the
Monitor interface.

Thanks to the factory inerface and the Monitor interface, Monitor
objects can still be mocked from lower-level components, without
mentioning the name of a concrete implementation of the Monitor
interface.

PS: the monitor class has also been divided into a MonitorData and
MonitorControl class.