Add isLoggedIn() to Monitor interface
Implemenatation of this method calls isLoggedIn() from Telegram. That
is, the is_logged_in.py Telethon script is executed. In this way, if at
application startup there is already a logged-in master.session file
(e.g. after a system reboot), this logged-in state can be reused.

This should still be implemented, that the application can detect such a
situation and resume operation without requiring a new login procedure.